Reorganizing all project
