+++ .dockerignore
@@ -1,0 +1,9 @@
+.git
+.ipynb_checkpoints/*
+Dockerfile
+.DS_Store
+.gitignore
+*/README.md
+env.*
+*.env.*
+*/venv/*
