--- grouping/exceptions.py
@@ -1,6 +1,0 @@
-class BadRequestException(Exception):
-    pass
-
-
-class UnknownException(Exception):
-    pass
